requirements.txt
flask
python-binance
python-dotenv
psycopg2
cryptography
requests
